separated common character from base character.
